add type2の呼び出し/計算で座標を出してわたす

+++ server/usecase/enemyUsecase.ts
@@ -7,14 +7,72 @@ import {
 import type { EnemyModel, EnemyModelWithPos } from '$/commonTypesWithClient/models';
 import { enemyRepository } from '$/repository/enemyRepository';
 import { computePosition } from '$/service/computePositions';
+import { type2EnemyMove } from '$/service/type2EnemyMove';
 import { entityChangeWithPos } from '$/service/entityChangeWithPos';
 import { enemyIdParser } from '$/service/idParsers';
 import { randomUUID } from 'crypto';
 
 const updateIntervalId: NodeJS.Timeout[] = [];
 const createIntervalId: NodeJS.Timeout[] = [];
+
 const ENEMY_UPDATE_INTERVAL = 100;
 const ENEMY_CREATE_INTERVAL = 1000;
+
+const TYPE0_ENEMY_LIMIT = 12;
+const TYPE1_ENEMY_LIMIT = 6;
+const TYPE2_ENEMY_LIMIT = 6;
+
+type Count = [number, number, number];
+
+const createStrongEnemies = async (counts: Count[]) => {
+  await Promise.all(
+    counts.map(
+      (count, displayNum) =>
+        //  [
+        // Promise.all(
+        //   [...Array(TYPE1_ENEMY_LIMIT - count[1])].map(async () => {
+        //     return
+        enemyRepository
+          .create({
+            id: enemyIdParser.parse(randomUUID()),
+            direction: {
+              x: 0,
+              y: 0,
+            },
+            createdPos: {
+              x: 1835 + SCREEN_WIDTH * displayNum,
+              y: Math.random() >= 0.5 ? 0 : 540,
+            },
+            createdAt: Date.now(),
+            type: 1,
+          })
+          .catch((error) => console.error(error))
+      // }
+    )
+    // ),
+    // Promise.all(
+    //   [...Array(TYPE2_ENEMY_LIMIT - count[2])].map(async () => {
+    //     return enemyRepository
+    //       .create({
+    //         id: enemyIdParser.parse(randomUUID()),
+    //         direction: {
+    //           x: 0,
+    //           y: 0,
+    //         },
+    //         createdPos: {
+    //           x: 1835 + SCREEN_WIDTH * displayNum,
+    //           y: Math.random() >= 0.5 ? 0 : 540,
+    //         },
+    //         createdAt: Date.now(),
+    //         type: 2,
+    //       })
+    //       .catch((error) => console.error(error));
+    //   })
+    // ),
+    // ])
+    // .flat(1)
+  );
+};
 
 export const enemyUseCase = {
   init: () => {
@@ -43,18 +101,21 @@ export const enemyUseCase = {
   create: async (): Promise<EnemyModel | null> => {
     const enemies = await enemyRepository.findAll();
 
-    const count: [number, number, number] = [0, 0, 0];
+    const counts: Count[] = [...Array(DISPLAY_COUNT)].map(() => [0, 0, 0]);
 
-    const countByType = () => {
+    const countsByType = () => {
       enemies.forEach((enemy) => {
-        if (count[enemy.type - 1] === undefined) return;
-        count[enemy.type]++;
+        const displayNum = Math.floor(enemy.createdPos.x / SCREEN_WIDTH);
+        if (counts[displayNum][enemy.type - 1] === undefined) return;
+        counts[displayNum][enemy.type]++;
       });
     };
 
-    countByType();
+    countsByType();
 
-    if (count[0] > 12) return null;
+    createStrongEnemies(counts);
+
+    if (counts.map((count) => count[0]).reduce((a, b) => a + b) > TYPE0_ENEMY_LIMIT) return null;
 
     const newEnemy = await enemyRepository.create({
       id: enemyIdParser.parse(randomUUID()),
@@ -90,7 +151,7 @@ export const enemyUseCase = {
     await Promise.all(
       currentEnemyList.map((enemy) => {
         const pos = computePosition(enemy);
-        if (outOfDisplay(pos)) {
+        if (outOfDisplay(pos) && enemy.type === 0) {
           return enemyRepository.delete(enemy.id);
         }
       })
@@ -100,7 +161,7 @@ export const enemyUseCase = {
   getEnemiesByDisplay: async (displayNumber: number): Promise<EnemyModelWithPos[]> => {
     const enemies = await enemyRepository.findAll();
     const getEnemiesByDisplay = enemies
-      .map(entityChangeWithPos)
+      .map((enemy) => (enemy.type === 0 ? entityChangeWithPos(enemy) : type2EnemyMove(enemy)))
       .filter(
         (enemy) =>
           enemy.pos.x + ENEMY_HALF_WIDTH > SCREEN_WIDTH * displayNumber &&
@@ -108,5 +169,14 @@ export const enemyUseCase = {
       ) as EnemyModelWithPos[];
 
     return getEnemiesByDisplay;
+  },
+
+  getEnemiesAll: async () => {
+    const enemies = await enemyRepository.findAll();
+    const enemiesWithPos = enemies.map((enemy) =>
+      enemy.type === 0 ? entityChangeWithPos(enemy) : type2EnemyMove(enemy)
+    );
+
+    return enemiesWithPos as EnemyModelWithPos[];
   },
 };
